Correct Javascript Transcription Error
The line:
```C
ndivs = (int)(fabsf(da) / (NSVG_PI*0.5f) + 1.0f);
```
Specifically includes the casting of ndivs to `int` type. This is not a throwaway form change since ndivs is the number-of-divisions and needs to be a discrete value. The casting of `float` to `int` in C is a floor function.